# granja-corales-main/SolutionBackEnd/Solution.API/bin/Release/netcoreapp3.1/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Solution.API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:62718"
    }
  ],
  "paths": {
...
    "/api/Articulos/{id}": {
      "get": {
        "tags": [
          "Articulos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Articulos"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Articulos"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Articulos"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Articulos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Articulos"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Articulos"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Articulos"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Articulos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Articulos"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Articulos"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Articulos"
                }
              }
            }
          }
        }
      }
    },
    "/api/ArticulosSolicitud": {
      "get": {
        "tags": [
          "ArticulosSolicitud"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticulosSolicitud"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticulosSolicitud"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticulosSolicitud"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ArticulosSolicitud"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticulosSolicitud"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticulosSolicitud"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ArticulosSolicitud"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ArticulosSolicitud"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticulosSolicitud"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticulosSolicitud"
                }
              }
            }
          }
        }
      }
    },
    "/api/ArticulosSolicitud/{id}": {
      "get": {
        "tags": [
          "ArticulosSolicitud"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ArticulosSolicitud"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticulosSolicitud"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticulosSolicitud"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ArticulosSolicitud"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticulosSolicitud"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticulosSolicitud"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ArticulosSolicitud"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ArticulosSolicitud"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ArticulosSolicitud"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticulosSolicitud"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticulosSolicitud"
                }
              }
            }
          }
        }
      }
    },
    "/api/Categorias": {
      "get": {
        "tags": [
          "Categorias"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Categorias"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Categorias"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Categorias"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categorias"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Categorias"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Categorias"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Categorias"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Categorias"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categorias"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categorias"
                }
              }
            }
          }
        }
      }
    },
    "/api/Categorias/{id}": {
      "get": {
        "tags": [
          "Categorias"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Categorias"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categorias"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categorias"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Categorias"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Categorias"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Categorias"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Categorias"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Categorias"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Categorias"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categorias"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Categorias"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Articulos": {
        "type": "object",
        "properties": {
          "idArticulo": {
            "type": "integer",
            "format": "int32"
          },
          "nombreCientifico": {
            "type": "string",
            "nullable": true
          },
          "familia": {
            "type": "string",
            "nullable": true
          },
          "tipo": {
            "type": "string",
            "nullable": true
          },
          "nombreComun": {
            "type": "string",
            "nullable": true
          },
          "dificultad": {
            "type": "string",
            "nullable": true
          },
          "temperamento": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "dieta": {
            "type": "string",
            "nullable": true
          },
          "tamanoMax": {
            "type": "string",
            "nullable": true
          },
          "origen": {
            "type": "string",
            "nullable": true
          },
          "tamanoMinPecera": {
            "type": "string",
            "nullable": true
          },
          "modificadoPor": {
            "type": "string",
            "nullable": true
          },
          "fecModificacion": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "imagePath": {
            "type": "string",
            "nullable": true
          },
          "idCategoria": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idCategoriaNavigation": {
            "$ref": "#/components/schemas/Categorias"
          }
        },
        "additionalProperties": false
      },
      "ArticulosSolicitud": {
        "type": "object",
        "properties": {
          "idArticuloSolicitud": {
            "type": "integer",
            "format": "int32"
          },
          "idSolicitud": {
            "type": "integer",
            "format": "int32"
          },
          "idArticulo": {
            "type": "integer",
            "format": "int32"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idArticuloNavigation": {
            "$ref": "#/components/schemas/Articulos"
          },
          "idSolicitudNavigation": {
            "$ref": "#/components/schemas/Solicitudes"
          }
        },
        "additionalProperties": false
      },
      "Categorias": {
        "type": "object",
        "properties": {
          "idCategoria": {
            "type": "integer",
            "format": "int32"
          },
          "tipo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Solicitudes": {
        "type": "object",
        "properties": {
          "idSolicitud": {
            "type": "integer",
            "format": "int32"
          },
          "fecCreacion": {
            "type": "string",
            "format": "date-time"
          },
          "estadoSolicitud": {
            "type": "string",
            "nullable": true
          },
          "idUsuario": {
            "type": "string",
            "nullable": true
          },
          "estadoAprobacion": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}